+++ src/include/logging/dummy_log_manager.h
@@ -35,11 +35,7 @@ public:
 
   virtual const std::string GetDirectories() final { return dirs_; }
 
-  // Worker side logic
-  virtual void RegisterWorker(size_t thread_id UNUSED_ATTRIBUTE) final {};
-  virtual void DeregisterWorker() final {};
-
-  virtual void DoRecovery(const size_t &begin_eid UNUSED_ATTRIBUTE) final {};
+  virtual void DoRecovery() final {};
 
   virtual void StartLoggers() final {};
   virtual void StopLoggers() final {};
